Convert to float before scale

+++ ultralytics/utils/plotting.py
@@ -393,8 +393,8 @@ class Annotator:
 
             ih, iw = self.im.shape[:2]
             if not retina_masks:
-                # Use scale_masks to properly remove padding and upsample
-                masks = ops.scale_masks(masks[None], (ih, iw))[0] > 0.5
+                # Use scale_masks to properly remove padding and upsample, convert bool to float first
+                masks = ops.scale_masks(masks[None].float(), (ih, iw))[0] > 0.5
                 # Convert original BGR image to RGB tensor
                 im_gpu = (
                     torch.from_numpy(self.im).to(masks.device).permute(2, 0, 1).flip(0).contiguous().float() / 255.0
